Markups.

+++ felix/fv/workload/workload.go
@@ -98,32 +98,36 @@ const defaultMTU = 1440 /* wiregueard mtu */
 func (w *Workload) Stop() {
 	if w == nil {
 		log.Info("Stop no-op because nil workload")
-	} else {
-		log.WithField("workload", w.Name).Info("Stop")
-		_ = w.C.ExecMayFail("sh", "-c", fmt.Sprintf("kill -9 %s & ip link del %s & ip netns del %s & wait", w.pid, w.InterfaceName, w.NamespaceID()))
-		// Killing the process inside the container should cause our long-running
-		// docker exec command to exit.  Do the Wait on a background goroutine,
-		// so we can time it out, just in case.
-		waitDone := make(chan struct{})
-		go func() {
-			defer close(waitDone)
-			_, err := w.runCmd.Process.Wait()
-			if err != nil {
-				log.WithField("workload", w.Name).Error("Failed to wait for docker exec, attempting to kill it.")
-				_ = w.runCmd.Process.Kill()
-			}
-		}()
-
-		select {
-		case <-waitDone:
-			log.WithField("workload", w.Name).Info("Workload stopped")
-		case <-time.After(10 * time.Second):
-			log.WithField("workload", w.Name).Error("Workload docker exec failed to exit?  Killing it.")
+		return
+	}
+	log.WithField("workload", w.Name).Info("Stop")
+	if !w.isRunning {
+		log.WithField("workload", w.Name).Info("Workload already stopped")
+		return
+	}
+	_ = w.C.ExecMayFail("sh", "-c", fmt.Sprintf("kill -9 %s & ip link del %s & ip netns del %s & wait", w.pid, w.InterfaceName, w.NamespaceID()))
+	// Killing the process inside the container should cause our long-running
+	// docker exec command to exit.  Do the Wait on a background goroutine,
+	// so we can time it out, just in case.
+	waitDone := make(chan struct{})
+	go func() {
+		defer close(waitDone)
+		_, err := w.runCmd.Process.Wait()
+		if err != nil {
+			log.WithField("workload", w.Name).Error("Failed to wait for docker exec, attempting to kill it.")
 			_ = w.runCmd.Process.Kill()
 		}
+	}()
 
-		w.isRunning = false
+	select {
+	case <-waitDone:
+		log.WithField("workload", w.Name).Info("Workload stopped")
+	case <-time.After(10 * time.Second):
+		log.WithField("workload", w.Name).Error("Workload docker exec failed to exit?  Killing it.")
+		_ = w.runCmd.Process.Kill()
 	}
+
+	w.isRunning = false
 }
 
 func Run(c *infrastructure.Felix, name, profile, ip, ports, protocol string, opts ...Opt) (w *Workload) {
